// apps/ripscloud-web/src/api/api.ts
export interface RipsAdminApplicationDTOsClientDto {
    'id'?: string;
    'tenantId'?: string;
    'nit'?: string;
    'verificationDigit'?: string;
    'companyName'?: string;
    'commercialName'?: string;
    'taxRegime'?: string;
    'economicActivityCode'?: string;
    'address'?: string;
    'departmentCode'?: string;
    'municipalityCode'?: string;
    'phoneNumber'?: string;
    'email'?: string;
    'typeOrganizationId'?: number;
    'typeDocumentIdentificationId'?: number;
    'isActive'?: boolean;
    'createdAt'?: string;
    'updatedAt'?: string | null;
}
// ...
export interface RipsAdminApplicationDTOsCreateClientRequestDto {
    'nit'?: string;
    'verificationDigit'?: string;
    'companyName'?: string;
    'commercialName'?: string;
    'taxRegime'?: string;
    'economicActivityCode'?: string;
    'address'?: string;
    'departmentCode'?: string;
    'municipalityCode'?: string;
    'phoneNumber'?: string;
    'email'?: string;
    'typeOrganizationId'?: number;
    'typeDocumentIdentificationId'?: number;
    'isActive'?: boolean;
}
// ...
export interface RipsAdminApplicationDTOsUpdateClientRequestDto {
    'nit'?: string;
    'verificationDigit'?: string;
    'companyName'?: string;
    'commercialName'?: string;
    'taxRegime'?: string;
    'economicActivityCode'?: string;
    'address'?: string;
    'departmentCode'?: string;
    'municipalityCode'?: string;
    'phoneNumber'?: string;
    'email'?: string;
    'typeOrganizationId'?: number;
    'typeDocumentIdentificationId'?: number;
    'isActive'?: boolean;
}